# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API sample",
    "version": "1.0.0",
    "description": "This is a sample API.",
    "contact": {
      "email": "admin@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://example.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "develop"
    },
    {
      "url": "https://example.com/api",
      "description": "production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "headers": {
      "x-total-count": {
        "example": 123,
        "type": "integer",
        "description": "Total number of items in the collection."
      }
    },
...
  },
  "paths": {
...
    "/posts": {
      "post": {
        "summary": "create a post",
        "description": "create a post",
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "post title",
                    "example": "YLorem ipsum dolor sit amet"
                  },
                  "content": {
                    "type": "string",
                    "description": "post content",
                    "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
                  },
                  "authorId": {
                    "type": "string",
                    "description": "author id",
                    "example": "clqj40goe0000z21ibpk26izw"
                  }
                },
                "required": [
                  "title",
                  "content",
                  "authorId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "post id",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "description": "post title",
                      "example": "YLorem ipsum dolor sit amet"
                    },
                    "content": {
                      "type": "string",
                      "description": "post content",
                      "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
                    },
                    "authorId": {
                      "type": "string",
                      "description": "author id",
                      "example": "clqj40goe0000z21ibpk26izw"
                    },
                    "createdAt": {
                      "type": "string",
                      "nullable": true,
                      "description": "created date",
                      "example": "2021-11-23T07:00:00Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "nullable": true,
                      "description": "updated date",
                      "example": "2021-11-23T07:00:00Z"
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "content",
                    "authorId",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Bad Request"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Internal Server Error"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "get posts list",
        "description": "get posts list",
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "description": "page number",
              "example": 1
            },
            "required": false,
            "description": "page number",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "default": 100,
              "description": "limit to take"
            },
            "required": false,
            "description": "limit to take",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "join key:direction with comma",
              "example": "id:desc,updatedAt:asc"
            },
            "required": false,
            "description": "join key:direction with comma",
            "name": "orderby",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "headers": {
              "x-total-count": {
                "$ref": "#/components/headers/x-total-count"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "post id",
                        "example": 1
                      },
                      "title": {
                        "type": "string",
                        "description": "post title",
                        "example": "YLorem ipsum dolor sit amet"
                      },
                      "content": {
                        "type": "string",
                        "description": "post content",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
                      },
                      "authorId": {
                        "type": "string",
                        "description": "author id",
                        "example": "clqj40goe0000z21ibpk26izw"
                      },
                      "createdAt": {
                        "type": "string",
                        "nullable": true,
                        "description": "created date",
                        "example": "2021-11-23T07:00:00Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "nullable": true,
                        "description": "updated date",
                        "example": "2021-11-23T07:00:00Z"
                      },
                      "author": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "user id",
                            "example": "clqj40goe0000z21ibpk26izw"
                          },
                          "userName": {
                            "type": "string",
                            "description": "unique user name",
                            "example": "John Doe"
                          },
                          "imageUrl": {
                            "type": "string",
                            "nullable": true,
                            "description": "user image",
                            "example": "/images/image.png"
                          }
                        },
                        "required": [
                          "id",
                          "userName",
                          "imageUrl"
                        ]
                      },
                      "bookmarks": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "userId": {
                              "type": "string",
                              "description": "user id",
                              "example": "clqj40goe0000z21ibpk26izw"
                            },
                            "user": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "description": "user id",
                                  "example": "clqj40goe0000z21ibpk26izw"
                                },
                                "userName": {
                                  "type": "string",
                                  "description": "unique user name",
                                  "example": "John Doe"
                                },
                                "imageUrl": {
                                  "type": "string",
                                  "nullable": true,
                                  "description": "user image",
                                  "example": "/images/image.png"
                                }
                              },
                              "required": [
                                "id",
                                "userName",
                                "imageUrl"
                              ]
                            }
                          },
                          "required": [
                            "userId",
                            "user"
                          ]
                        }
                      },
                      "_count": {
                        "type": "object",
                        "properties": {
                          "bookmarks": {
                            "type": "integer",
                            "example": 1
                          }
                        },
                        "required": [
                          "bookmarks"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "title",
                      "content",
                      "authorId",
                      "createdAt",
                      "updatedAt",
                      "author",
                      "bookmarks",
                      "_count"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Bad Request"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Not Found"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Internal Server Error"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}